requirements/production.txt
django==1.6

greenlet==0.4.2

gevent==1.0

gevent-socketio==0.3.6

redis==2.8.0
django-redis-sessions==0.4.0